fix(pr): resolve clang-tidy and system-tests CI failures
- structural_equal: cast InlineLanguage to unsigned int in error
  message (uint8_t passed to operator<< prints as char and clang-tidy
  flags bugprone-unintended-char-ostream-output).
- test_l3_parallel_reduce: drop 'self' from HOST SubWorker
  ``reduce_sum`` so it parses under the new self-contained form.

+++ src/ir/transforms/structural_equal.cpp
@@ -342,8 +342,8 @@ class StructuralEqualImpl {
     if (lhs != rhs) {
       if constexpr (AssertMode) {
         std::ostringstream msg;
-        msg << "InlineLanguage mismatch (" << static_cast<uint8_t>(lhs) << " != " << static_cast<uint8_t>(rhs)
-            << ")";
+        msg << "InlineLanguage mismatch (" << static_cast<unsigned int>(lhs)
+            << " != " << static_cast<unsigned int>(rhs) << ")";
         ThrowMismatch(msg.str(), IRNodePtr(), IRNodePtr(), "", "");
       }
       return false;
